requirements.txt
fastapi == 0.115.12
uvicorn == 0.34.3
mangum == 0.19.0
boto3 == 1.38.17
botocore == 1.38.17
python-dotenv == 1.1.0
requests == 2.32.3
pyjwt == 2.10.1
names == 0.3.0
pytz == 2024.2
pydantic == 2.11.4